+++ R/convert.R
@@ -71,12 +71,7 @@ scribe_convert <- function(method = c("default", "evaluate", "none")) {
     method,
     none = identity,
     default = value_convert,
-    evaluate = function(x, ...) {
-      if (!is.character(x)) {
-        return(x)
-      }
-      eval(str2expression(as.character(x)), baseenv())
-    }
+    evaluate = function(x, ...) eval(str2expression(as.character(x)), baseenv())
   )
 }